+++ sdk/identity/identity/src/credentials/managedIdentityCredential/cloudShellMsi.ts
@@ -19,8 +19,7 @@ const logger = credentialLogger(msiName);
  */
 function prepareRequestOptions(
   scopes: string | string[],
-  clientId?: string,
-  resourceId?: string
+  clientId?: string
 ): PipelineRequestOptions {
   const resource = mapScopesToResource(scopes);
   if (!resource) {
@@ -33,9 +32,6 @@ function prepareRequestOptions(
 
   if (clientId) {
     body.client_id = clientId;
-  }
-  if (resourceId) {
-    body.mi_res_id = resourceId;
   }
 
   // This error should not bubble up, since we verify that this environment variable is defined in the isAvailable() method defined below.
@@ -59,7 +55,13 @@ function prepareRequestOptions(
  * Defines how to determine whether the Azure Cloud Shell MSI is available, and also how to retrieve a token from the Azure Cloud Shell MSI.
  */
 export const cloudShellMsi: MSI = {
-  async isAvailable({ scopes }): Promise<boolean> {
+  async isAvailable({ scopes, resourceId }): Promise<boolean> {
+    if (resourceId) {
+      logger.info(
+        `${msiName}: Unavailable. User defined managed Identity by resource Id is not supported by Cloud Shell Managed Identity Endpoint.`
+      );
+      return false;
+    }
     const resource = mapScopesToResource(scopes);
     if (!resource) {
       logger.info(`${msiName}: Unavailable. Multiple scopes are not supported.`);
@@ -75,7 +77,7 @@ export const cloudShellMsi: MSI = {
     configuration: MSIConfiguration,
     getTokenOptions: GetTokenOptions = {}
   ): Promise<AccessToken | null> {
-    const { identityClient, scopes, clientId, resourceId } = configuration;
+    const { identityClient, scopes, clientId } = configuration;
 
     logger.info(
       `${msiName}: Using the endpoint coming form the environment variable MSI_ENDPOINT = ${process.env.MSI_ENDPOINT}.`
@@ -83,7 +85,7 @@ export const cloudShellMsi: MSI = {
 
     const request = createPipelineRequest({
       abortSignal: getTokenOptions.abortSignal,
-      ...prepareRequestOptions(scopes, clientId, resourceId),
+      ...prepareRequestOptions(scopes, clientId),
       // Generally, MSI endpoints use the HTTP protocol, without transport layer security (TLS).
       allowInsecureConnection: true,
     });